+++ bundles/analysis/analyse/view/StartAnalyse.js
@@ -2788,15 +2788,20 @@ Oskari.clazz.define('Oskari.analysis.bundle.analyse.view.StartAnalyse',
                 requestBuilder,
                 showFeatureDataReqBuilder,
                 request;
-
+                
             if (me._showFeatureDataWithoutSaving) {
                 var aggregateValues = analyseJson.aggregate,
                     geojson = analyseJson.geojson.features[0];
 
                 var rn = 'MapModulePlugin.AddFeaturesToMapRequest',
                     style = OpenLayers.Util.applyDefaults(style, OpenLayers.Feature.Vector.style['default']);
-
-                me.instance.sandbox.postRequestByName(rn, [geojson, 'GeoJSON', null, null, 'replace', true, style, false]);
+                me.instance.sandbox.postRequestByName(rn, [geojson, {
+                    layerId: 'ANALYSIS_VECTOR', 
+                    replace: 'replace',
+                    layerOptions: null,
+                    centerTo: false,
+                    featureStyle: style
+                }]);
 
                 me._showAggregateResultPopup(aggregateValues, geojson, noDataCnt);
             } else {